Split NativeDatabase into focused files
Break up NativeDatabase.swift into separate files per protocol conformance (NativeRecordReader, NativeRecordWriter, NativeRecordLocator, NativeActivityReader), move MatrixAggregator protocol and MatrixUploader into the Syncable folder, and consolidate ActivityPoint matrix logic.
