+++ Sources/Scout/Native/Database/NativeActivityReader.swift
@@ -5,7 +5,24 @@
 // license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
-import Foundation
+import CloudKit
+
+extension CKDatabase: ActivityReader {
+    func activity(in range: Range<Date>) async throws -> [ActivityPoint] {
+        let query = RecordQuery(
+            recordType: PeriodMatrix.self,
+            filters: range.dateFilters
+        )
+
+        let matrices = try await readAll(
+            matching: query,
+            fields: nil
+        )
+        .map(PeriodMatrix.init)
+
+        return ActivityPoint.points(from: matrices)
+    }
+}
 
 extension ActivityPoint {
     static func points(from matrices: [PeriodMatrix]) -> [ActivityPoint] {
