+++ Sources/Scout/Native/Matrix/Syncable/MatrixUploader.swift
@@ -1,11 +1,11 @@
 //
-// Copyright 2024 Mikhail Kasianov
+// Copyright 2026 Mikhail Kasianov
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
-import CloudKit
+import Foundation
 
 struct MatrixUploader<T: CellProtocol>: Sendable {
     let database: RecordWriter & RecordReader
